Move dependencies example to workflows folder and add tests
Updates sample to use relative imports.

Adds license headers.

Uses 4 spaces for indentation.
